matriculas/requirements.txt
Flask==3.1.0
mysql-connector-python==9.3.0
python-dotenv==1.1.0
Flask-Mail==0.10.0
Werkzeug==3.1.3
